+++ lib/src/ui/flutter_defender_message_id.dart
@@ -3,6 +3,7 @@ enum FlutterDefenderMessageId {
   screenshotsBlocked,
   overlaysBlocked,
   screenCaptureBlocked,
+  protectedContentHidden,
   foregroundRequired,
   rootOrJailbreakBlocked,
   proxyOrVpnBlocked,